requirements.txt
asgiref==3.8.1
Django==5.1.3
python-dotenv==1.0.1
sqlparse==0.5.1
tzdata==2024.2
Pillow==10.4.0
django-debug-toolbar==4.4.6
certifi~=2024.8.30
urllib3~=2.2.3
idna~=3.10
charset-normalizer~=3.3.2
filelock~=3.16.1
colorama~=0.4.6
chardet~=5.2.0
djangorestframework~=3.15.2